+++ src/pool_inner.rs
@@ -229,10 +229,9 @@ impl<M: Manager> PoolLockFuture<'_, M> {
     fn spawn_idle_conn(&self, inner: &mut MutexGuard<'_, PoolInner<M>>) {
         let shared = self.shared_pool;
         if inner.total() < shared.statics.max_size {
-            inner.incr_pending_inner(1);
             let shared_clone = shared.clone();
             shared.spawn(async move { shared_clone.add_idle_conn().await });
-            //                .unwrap_or_else(|_| inner.decr_pending_inner(1));
+            inner.incr_pending_inner(1);
         }
     }